fix:[FutebolPage]: refactoring
