+++ src/sections/@dashboard/app/AppChipMercados/Context/index.js
@@ -1,0 +1,3 @@
+import { createContext } from 'react';
+
+export const MarketContext = createContext("");
