+++ src/sections/@dashboard/app/AppChipMercados/components/FilterHoursButtons/constants.js
@@ -1,0 +1,36 @@
+const NEXT_HOURS_BUTTONS = [
+  {
+    title: '3 horas',
+    md: 3,
+    tab: 1,
+  },
+  {
+    title: '6 horas',
+    md: 3,
+    tab: 1,
+  },
+  {
+    title: '12 horas',
+    md: 3,
+    tab: 1,
+  },
+  {
+    title: '24 horas',
+    md: 3,
+    tab: 1,
+  },
+  {
+    title: 'Horário de início',
+    md: 4,
+    tab: 1,
+  },
+  {
+    title: 'Por competição',
+    md: 4,
+    tab: 1,
+  },
+];
+
+export {
+  NEXT_HOURS_BUTTONS
+}
